refactor: based on discussion
