+++ test/RelayClient.test.ts
@@ -3,8 +3,6 @@ import {
   DeployVerifier,
   DeployVerifier__factory,
   EnvelopingTypes,
-  IERC20,
-  IERC20__factory,
   IForwarder,
   IForwarder__factory,
   ISmartWalletFactory,
@@ -18,7 +16,6 @@ import { expect, use } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import { BigNumber, BigNumberish, constants, providers, Transaction, Wallet, utils, errors } from 'ethers';
 
-import * as etherUtils from '@ethersproject/transactions';
 import type { TransactionResponse, TransactionReceipt } from '@ethersproject/providers';
 import { solidityKeccak256 } from 'ethers/lib/utils';
 import Sinon, { SinonStub, SinonStubbedInstance } from 'sinon';
@@ -34,13 +31,11 @@ import type {
   RelayRequest,
   UserDefinedEnvelopingRequest,
   UserDefinedRelayData,
-  UserDefinedRelayRequest,
   UserDefinedRelayRequestBody,
 } from '../src/common/relayRequest.types';
 import type { EnvelopingTxRequest } from '../src/common/relayTransaction.types';
 import {
   MISSING_CALL_FORWARDER,
-  MISSING_SMART_WALLET_ADDRESS,
   NOT_RELAYED_TRANSACTION,
 } from '../src/constants/errorMessages';
 import EnvelopingEventEmitter, {
@@ -48,7 +43,6 @@ import EnvelopingEventEmitter, {
 } from '../src/events/EnvelopingEventEmitter';
 import RelayClient, {
 } from '../src/RelayClient';
-import { INTERNAL_TRANSACTION_ESTIMATED_CORRECTION } from '../src/utils';
 import { FAKE_ENVELOPING_CONFIG } from './config.fakes';
 import { FAKE_HUB_INFO } from './relayHub.fakes';
 import {
@@ -57,14 +51,12 @@ import {
   FAKE_ENVELOPING_REQUEST_DATA,
   FAKE_RELAY_REQUEST,
   FAKE_RELAY_REQUEST_BODY,
-  FAKE_RELAY_TRANSACTION_REQUEST,
-  FAKE_REQUEST_CONFIG
+  FAKE_RELAY_TRANSACTION_REQUEST
 } from './request.fakes';
-import type { TokenGasEstimationParams, RequestConfig } from '../src/common/relayClient.types';
 
+import * as etherUtils from '@ethersproject/transactions';
 import * as clientConfiguration from '../src/common/clientConfigurator';
 import * as relayUtils from '../src/utils';
-import * as discoveryUtils from '../src/discovery/utils';
 import * as gasEstimator from '../src/gasEstimator/gasEstimator';
 
 
@@ -89,34 +81,32 @@ describe('RelayClient', function () {
   });
 
   describe('constructor', function () {
-    it('should store provider', function () {
-      const expectedProvider: providers.BaseProvider =
-        sandbox.createStubInstance(providers.JsonRpcProvider);
+    it('should store httpClient', function () {
+      const expectedHttpClient: HttpClient =
+        sandbox.createStubInstance(HttpClient);
       const relayClient = new RelayClient();
 
-      (relayClient as unknown as { _provider: providers.Provider })._provider =
-        expectedProvider;
+      (relayClient as unknown as { _httpClient: HttpClient })._httpClient =
+      expectedHttpClient;
 
       expect(relayClient).to.be.instanceOf(RelayClient);
       expect(
-        (relayClient as unknown as { _provider: providers.Provider })._provider,
+        (relayClient as unknown as { _httpClient: HttpClient })._httpClient,
         'Provider'
-      ).to.equal(expectedProvider);
+      ).to.equal(expectedHttpClient);
     });
   });
 
   describe('methods', function () {
     type RelayClientExposed = {
       _envelopingConfig: EnvelopingConfig;
-      _provider: providers.Provider;
       _httpClient: HttpClient;
       _prepareHttpRequest: (
         _hubInfo: HubInfo,
         request: EnvelopingTypes.RelayRequestStruct
       ) => Promise<EnvelopingTxRequest>;
       _getEnvelopingRequestDetails: (
-        request: UserDefinedEnvelopingRequest,
-        requestConfig: RequestConfig
+        request: UserDefinedEnvelopingRequest
       ) => Promise<EnvelopingRequest>;
       _calculateGasPrice: () => Promise<BigNumber>;
       _attemptRelayTransaction: (relayInfo: RelayInfo, envelopingTx: EnvelopingTxRequest) => Promise<Transaction>;
@@ -133,10 +123,11 @@ describe('RelayClient', function () {
     let forwarderStub: Sinon.SinonStubbedInstance<IForwarder>;
     let factoryStub: Sinon.SinonStubbedInstance<ISmartWalletFactory>;
     let ethersLogger: utils.Logger;
+    let providerStub: providers.Provider;
 
     beforeEach(function () {
       relayClient = new RelayClient() as unknown as RelayClientExposed;
-      relayClient._provider = {
+      providerStub = {
         _isProvider: true,
         getTransactionCount: sandbox.stub().resolves(FAKE_TX_COUNT),
         getNetwork: sandbox.stub().resolves({
@@ -144,6 +135,9 @@ describe('RelayClient', function () {
         } as Network),
         estimateGas: sandbox.stub().resolves(FAKE_GAS_PRICE),
       } as unknown as providers.Provider;
+
+      sandbox.stub(clientConfiguration, 'getProvider').returns(providerStub);
+      
 
       forwarderStub = {
         nonce: sandbox.stub(),
@@ -165,7 +159,7 @@ describe('RelayClient', function () {
 
       beforeEach(function () {
         fakeTokenGasEstimation = constants.Two;
-        relayClient.estimateTokenTransferGas = sandbox.stub().returns(fakeTokenGasEstimation);
+        sandbox.stub(relayUtils, 'estimateTokenTransferGas').returns(Promise.resolve(fakeTokenGasEstimation));
         accountManagerStub = {
           sign: sandbox
             .stub(AccountManager.prototype, 'sign')
@@ -292,7 +286,7 @@ describe('RelayClient', function () {
         };
 
         const fakeTokenGas = constants.Two;
-        sandbox.stub(RelayClient.prototype, 'estimateTokenTransferGas').returns(Promise.resolve(fakeTokenGas));
+        sandbox.stub(relayUtils, 'estimateTokenTransferGas').returns(Promise.resolve(fakeTokenGas));
         forwarderStub.nonce.resolves(constants.Two);
         const signStub = accountManagerStub.sign;
         await relayClient._prepareHttpRequest(
@@ -342,8 +336,7 @@ describe('RelayClient', function () {
               ...FAKE_ENVELOPING_REQUEST_DATA,
               callForwarder: undefined,
             },
-          } as unknown as UserDefinedEnvelopingRequest,
-          {}
+          } as unknown as UserDefinedEnvelopingRequest
         );
 
         await expect(call).to.be.eventually.rejected;
@@ -364,8 +357,7 @@ describe('RelayClient', function () {
               ...FAKE_ENVELOPING_REQUEST_DATA,
               callVerifier: undefined,
             },
-          },
-          {}
+          }
         );
 
         expect(actualCallVerifier).to.equal(expectedCallVerifier);
@@ -383,8 +375,7 @@ describe('RelayClient', function () {
               ...FAKE_ENVELOPING_REQUEST_DATA,
               callVerifier: undefined,
             },
-          },
-          {}
+          }
         );
 
         expect(actualCallVerifier).to.equal(expectedCallVerifier);
@@ -404,8 +395,7 @@ describe('RelayClient', function () {
               ...FAKE_ENVELOPING_REQUEST_DATA,
               callVerifier: undefined,
             },
-          },
-          {}
+          }
         );
 
         await expect(call).to.be.eventually.rejected;
@@ -426,8 +416,7 @@ describe('RelayClient', function () {
               ...FAKE_ENVELOPING_REQUEST_DATA,
               feesReceiver: undefined,
             },
-          },
-          {}
+          }
         );
 
         expect(actualFeesReceiver.toString()).to.equal(expectedFeesReceiver);
@@ -438,8 +427,7 @@ describe('RelayClient', function () {
         const {
           relayData: { gasPrice: actualGasPrice },
         } = await relayClient._getEnvelopingRequestDetails(
-          FAKE_RELAY_REQUEST,
-          {}
+          FAKE_RELAY_REQUEST
         );
 
         expect(actualGasPrice).to.equal(expectedGasPrice);
@@ -462,8 +450,7 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.relayData,
               gasPrice: undefined,
             },
-          },
-          {}
+          }
         );
 
         expect(actualGasPrice).to.equal(expectedGasPrice);
@@ -479,8 +466,7 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.relayData,
               gasPrice: undefined,
             },
-          },
-          {}
+          }
         );
 
         await expect(call).to.be.eventually.rejected;
@@ -497,8 +483,7 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.request,
               data: undefined,
             },
-          } as unknown as UserDefinedEnvelopingRequest,
-          {}
+          } as unknown as UserDefinedEnvelopingRequest
         );
 
         await expect(call).to.be.eventually.rejected;
@@ -515,8 +500,7 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.request,
               from: undefined,
             },
-          } as unknown as UserDefinedEnvelopingRequest,
-          {}
+          } as unknown as UserDefinedEnvelopingRequest
         );
 
         await expect(call).to.be.eventually.rejected;
@@ -533,8 +517,7 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.request,
               to: undefined,
             },
-          } as unknown as UserDefinedEnvelopingRequest,
-          {}
+          } as unknown as UserDefinedEnvelopingRequest
         );
 
         await expect(call).to.be.eventually.rejected;
@@ -555,8 +538,7 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.request,
               value: undefined,
             },
-          } as unknown as UserDefinedEnvelopingRequest,
-          {}
+          } as unknown as UserDefinedEnvelopingRequest
         );
 
         expect(actualValue).to.equal(expectedValue);
@@ -574,8 +556,7 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.request,
               tokenAmount: undefined,
             },
-          } as unknown as UserDefinedEnvelopingRequest,
-          {}
+          } as unknown as UserDefinedEnvelopingRequest
         );
 
         expect(actualTokenAmount).to.equal(expectedTokenAmount);
@@ -589,8 +570,7 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.request,
               tokenContract: undefined,
             },
-          } as unknown as UserDefinedEnvelopingRequest,
-          {}
+          } as unknown as UserDefinedEnvelopingRequest
         );
 
         await expect(call).to.be.eventually.rejected;
@@ -611,8 +591,7 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.request,
               nonce: undefined,
             },
-          } as unknown as EnvelopingRequest,
-          {}
+          } as unknown as EnvelopingRequest
         );
 
         expect(actualNonce).to.equal(expectedNonce);
@@ -630,8 +609,7 @@ describe('RelayClient', function () {
               ...FAKE_RELAY_REQUEST.request,
               nonce: undefined,
             },
-          } as unknown as EnvelopingRequest,
-          {}
+          } as unknown as EnvelopingRequest
         );
 
         expect(actualNonce).to.equal(expectedNonce);
@@ -643,8 +621,7 @@ describe('RelayClient', function () {
         const {
           request: { relayHub: actualRelayHub },
         } = await relayClient._getEnvelopingRequestDetails(
-          FAKE_DEPLOY_REQUEST,
-          {}
+          FAKE_DEPLOY_REQUEST
         );
 
         expect(actualRelayHub).to.equal(expectedRelayHubAddr);
@@ -662,8 +639,7 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.request,
               relayHub: undefined,
             },
-          },
-          {}
+          }
         );
 
         expect(actualRelayHub).to.equal(expectedRelayHubAddr);
@@ -682,8 +658,7 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.request,
               relayHub: undefined,
             },
-          },
-          {}
+          }
         );
 
         await expect(call).to.be.eventually.rejected;
@@ -696,8 +671,7 @@ describe('RelayClient', function () {
         const {
           request: { tokenGas: actualTokenGas },
         } = await relayClient._getEnvelopingRequestDetails(
-          FAKE_DEPLOY_REQUEST,
-          {}
+          FAKE_DEPLOY_REQUEST
         );
 
         expect(actualTokenGas).to.equal(constants.Zero);
@@ -710,37 +684,14 @@ describe('RelayClient', function () {
         const {
           request: { gas: actualGas },
         } = (await relayClient._getEnvelopingRequestDetails(
-          FAKE_RELAY_REQUEST,
-          {}
+          FAKE_RELAY_REQUEST
         )) as RelayRequest;
 
         expect(actualGas).to.equal(expectedGas);
       });
 
-      it('should return given forceGasLimit from request config in the request body for relay request', async function () {
-        const expectedGas = BigNumber.from((Math.random() * 100).toFixed(0));
-        const {
-          request: { gas: actualGas },
-        } = (await relayClient._getEnvelopingRequestDetails(
-          {
-            ...FAKE_RELAY_REQUEST,
-            request: {
-              ...FAKE_RELAY_REQUEST_BODY,
-              gas: undefined,
-            },
-          } as unknown as UserDefinedRelayRequest,
-          {
-            forceGasLimit: expectedGas.toString(),
-          }
-        )) as RelayRequest;
-
-        expect(actualGas.toString()).to.equal(expectedGas.toString());
-      });
-
-      it('should throw if no gas limit given to relay request or forceGasLimit to request config', async function () {
-        relayClient.estimateInternalCallGas = sandbox
-          .stub()
-          .resolves(undefined);
+      it('should throw if no gas limit given to relay request', async function () {
+        sandbox.stub(relayUtils, 'estimateInternalCallGas').resolves(undefined);
         const call = relayClient._getEnvelopingRequestDetails(
           {
             ...FAKE_RELAY_REQUEST,
@@ -748,8 +699,7 @@ describe('RelayClient', function () {
               ...FAKE_RELAY_REQUEST_BODY,
               gas: undefined,
             },
-          },
-          {}
+          }
         );
 
         await expect(call).to.be.eventually.rejected;
@@ -766,8 +716,7 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.request,
               gas: undefined,
             },
-          },
-          {}
+          }
         );
 
         await expect(call).not.to.be.eventually.rejectedWith(
@@ -780,8 +729,7 @@ describe('RelayClient', function () {
         const {
           request: { index: actualIndex },
         } = (await relayClient._getEnvelopingRequestDetails(
-          FAKE_DEPLOY_REQUEST,
-          {}
+          FAKE_DEPLOY_REQUEST
         )) as DeployRequest;
 
         expect(actualIndex).to.equal(expectedIndex);
@@ -798,8 +746,7 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.request,
               index: undefined,
             },
-          },
-          {}
+          }
         )) as DeployRequest;
 
         expect(actualIndex).to.equal(expectedIndex);
@@ -810,8 +757,7 @@ describe('RelayClient', function () {
         const {
           request: { recoverer: actualRecoverer },
         } = (await relayClient._getEnvelopingRequestDetails(
-          FAKE_DEPLOY_REQUEST,
-          {}
+          FAKE_DEPLOY_REQUEST
         )) as DeployRequest;
 
         expect(actualRecoverer).to.equal(expectedRecoverer);
@@ -828,110 +774,10 @@ describe('RelayClient', function () {
               ...FAKE_DEPLOY_REQUEST.request,
               recoverer: undefined,
             },
-          },
-          {}
+          }
         )) as DeployRequest;
 
         expect(actualRecoverer).to.equal(expectedRecoverer);
-      });
-    });
-
-    describe('estimateInternalCallGas', function () {
-      let relayRequest: EnvelopingTypes.RelayRequestStruct;
-
-      beforeEach(function () {
-        relayRequest = {
-          request: {
-            from: createRandomAddress(),
-            to: createRandomAddress(),
-            data: '0x01',
-          },
-          relayData: {
-            gasPrice: '60000000',
-          },
-        } as EnvelopingTypes.RelayRequestStruct;
-      });
-
-      it('should return the estimation applying the internal correction', async function () {
-        const estimateGas = BigNumber.from(10000);
-        relayClient._provider.estimateGas = sandbox
-          .stub()
-          .resolves(estimateGas);
-        const internalEstimationCorrection = 5000;
-        const estimation = await relayClient.estimateInternalCallGas({
-          data: relayRequest.request.data,
-          from: relayRequest.request.from,
-          to: relayRequest.request.to,
-          gasPrice: relayRequest.relayData.gasPrice,
-          internalEstimationCorrection,
-        });
-        const expectedEstimation = estimateGas.sub(
-          internalEstimationCorrection
-        );
-
-        expect(estimation.toString()).to.be.equal(
-          expectedEstimation.toString()
-        );
-      });
-
-      it('should return the estimation without applying the internal correction', async function () {
-        const expectedGasEst = BigNumber.from(4000);
-        relayClient._provider.estimateGas = sandbox
-          .stub()
-          .resolves(expectedGasEst);
-        const internalEstimationCorrection = 5000;
-        const actualGasEst = await relayClient.estimateInternalCallGas({
-          data: relayRequest.request.data,
-          from: relayRequest.request.from,
-          to: relayRequest.request.to,
-          gasPrice: relayRequest.relayData.gasPrice,
-          internalEstimationCorrection,
-        });
-
-        expect(actualGasEst.toString()).to.be.equal(expectedGasEst.toString());
-      });
-
-      it('should return estimation without applying gas correction factor when addExternalCorrection is set to false', async function () {
-        const estimateGas = BigNumber.from(10000);
-        relayClient._provider.estimateGas = sandbox
-          .stub()
-          .resolves(estimateGas);
-        const internalEstimationCorrection = 5000;
-        const estimation = await relayClient.estimateInternalCallGas({
-          data: relayRequest.request.data,
-          from: relayRequest.request.from,
-          to: relayRequest.request.to,
-          gasPrice: relayRequest.relayData.gasPrice,
-          internalEstimationCorrection,
-          estimatedGasCorrectionFactor: '1',
-        });
-        const expectedEstimation = estimateGas.sub(
-          internalEstimationCorrection
-        );
-
-        expect(estimation.toString()).to.be.equal(
-          expectedEstimation.toString()
-        );
-      });
-
-      it('should return estimation applying gas correction factor when estimatedGasCorrectionFactor different from 1', async function () {
-        const expectedEstimation = BigNumber.from(7_500);
-        relayClient._provider.estimateGas = sandbox
-          .stub()
-          .resolves(BigNumber.from(10_000));
-        const internalEstimationCorrection = 5_000;
-        const estimation = await relayClient.estimateInternalCallGas({
-          data: relayRequest.request.data,
-          from: relayRequest.request.from,
-          to: relayRequest.request.to,
-          gasPrice: relayRequest.relayData.gasPrice,
-          internalEstimationCorrection,
-          estimatedGasCorrectionFactor: '1.5',
-        });
-
-        expect(estimation.toString()).to.be.equal(
-          expectedEstimation.toString()
-        );
       });
     });
 
@@ -940,7 +786,7 @@ describe('RelayClient', function () {
       it('should return minGasPrice when minGasPrice is higher than gas price from provider', async function () {
         const expectedGasPrice = 30000;
         relayClient._envelopingConfig.minGasPrice = expectedGasPrice;
-        relayClient._provider.getGasPrice = sandbox
+        providerStub.getGasPrice = sandbox
           .stub()
           .callsFake(() => {
             console.log('getGasPrice called');
@@ -953,7 +799,7 @@ describe('RelayClient', function () {
 
       it('should return gas price multiplied by correction (factor + 1)', async function () {
         const estimateGas = BigNumber.from(60000);
-        relayClient._provider.getGasPrice = sandbox
+        providerStub.getGasPrice = sandbox
           .stub()
           .resolves(estimateGas);
         const expectedGasEstimate = estimateGas.mul(
@@ -1002,222 +848,7 @@ describe('RelayClient', function () {
       });
     });
 
-    describe('estimateTokenTransferGas', function () {
-      let relayClient: RelayClient;
-
-      let ierc20TransferStub: SinonStub;
-
-      beforeEach(function () {
-        relayClient = new RelayClient();
-
-        ierc20TransferStub = sandbox.stub();
-        const estimateGas = {
-          transfer: ierc20TransferStub,
-        };
-
-        const ierc20Stub = {
-          estimateGas,
-        } as unknown as Sinon.SinonStubbedInstance<IERC20>;
-
-        sandbox.stub(IERC20__factory, 'connect').returns(ierc20Stub);
-      });
-
-      it('should return 0 if token contract is zero address', async function () {
-        const request: TokenGasEstimationParams = {
-          relayRequest: {
-            ...FAKE_RELAY_REQUEST,
-            request: {
-              ...FAKE_RELAY_REQUEST.request,
-              tokenContract: constants.AddressZero,
-            }
-          },
-          ...FAKE_REQUEST_CONFIG
-        };
-
-        expect(
-          (await relayClient.estimateTokenTransferGas(request)).toString()
-        ).to.be.equal(BigNumber.from(0).toString());
-      });
-
-      it('should return 0 if token amount is 0', async function () {
-        const request: TokenGasEstimationParams = {
-          relayRequest: {
-            ...FAKE_RELAY_REQUEST,
-            request: {
-              ...FAKE_RELAY_REQUEST.request,
-              tokenAmount: constants.Zero,
-            }
-          },
-          ...FAKE_REQUEST_CONFIG
-        };
-
-        expect(
-          (await relayClient.estimateTokenTransferGas(request)).toString()
-        ).to.be.equal(constants.Zero.toString());
-      });
-
-      it('should fail if it is a deploy and the smartWallet is missing', async function () {
-        const request: TokenGasEstimationParams = {
-          relayRequest: {...FAKE_DEPLOY_REQUEST},
-          preDeploySWAddress: undefined,
-        };
-
-        await expect(
-          relayClient.estimateTokenTransferGas(request)
-        ).to.be.rejectedWith(MISSING_SMART_WALLET_ADDRESS);
-      });
-
-      it('should fail if it is a deploy and the smartWallet is the zero address', async function () {
-        const request: TokenGasEstimationParams = {
-          relayRequest: {...FAKE_DEPLOY_REQUEST},
-          preDeploySWAddress: constants.AddressZero,
-        };
-
-        await expect(
-          relayClient.estimateTokenTransferGas(request)
-        ).to.be.rejectedWith(MISSING_SMART_WALLET_ADDRESS);
-      });
-
-      it('should fail if it is a relay transaction and the callForwarder is missing', async function () {
-        const request: TokenGasEstimationParams = {
-          relayRequest: {
-            ...FAKE_RELAY_REQUEST,
-            relayData: {
-              ...FAKE_RELAY_REQUEST.relayData,
-              callForwarder: constants.AddressZero,
-            }
-          },
-          preDeploySWAddress: constants.AddressZero
-        };
-
-        await expect(
-          relayClient.estimateTokenTransferGas(request)
-        ).to.be.rejectedWith(MISSING_CALL_FORWARDER);
-      });
-
-      it('should correct the value of the estimation when the gas cost is greater than the correction', async function () {
-        const FAKE_GAS_COST = 30000;
-        const INTERNAL_CORRECTION = 20000;
-        const EXPECTED_ESTIMATION = FAKE_GAS_COST - INTERNAL_CORRECTION;
-
-        const request: TokenGasEstimationParams = {
-          relayRequest: {
-            ...FAKE_RELAY_REQUEST,
-            relayData: {
-              ...FAKE_RELAY_REQUEST.relayData,
-              gasPrice: FAKE_GAS_COST,
-            }
-          },
-          internalEstimationCorrection: INTERNAL_CORRECTION
-        };
-
-        ierc20TransferStub.resolves(BigNumber.from(FAKE_GAS_COST));
-
-        const estimation = await relayClient.estimateTokenTransferGas(request);
-
-        expect(estimation.toString()).to.equal(EXPECTED_ESTIMATION.toString());
-      });
-
-      it('should not correct the value of the estimation when the gas cost is lower than the correction', async function () {
-        const FAKE_GAS_COST = 10000;
-        const INTERNAL_CORRECTION = 20000;
-        const EXPECTED_ESTIMATION = FAKE_GAS_COST;
-
-        ierc20TransferStub.resolves(BigNumber.from(FAKE_GAS_COST));
-
-        const request: TokenGasEstimationParams = {
-          relayRequest: {
-            ...FAKE_RELAY_REQUEST,
-            relayData: {
-              ...FAKE_RELAY_REQUEST.relayData,
-              gasPrice: FAKE_GAS_COST,
-            }
-          },
-          internalEstimationCorrection: INTERNAL_CORRECTION
-        };
-
-        const estimation = await relayClient.estimateTokenTransferGas(request);
-
-        expect(estimation.toString()).to.equal(EXPECTED_ESTIMATION.toString());
-      });
-
-      it('should apply the correction factor', async function () {
-        const FAKE_GAS_COST = 10000;
-        const INTERNAL_CORRECTION = 20000;
-        const CORRECTION_FACTOR = 1.5;
-        const EXPECTED_ESTIMATION = FAKE_GAS_COST * CORRECTION_FACTOR;
-
-        ierc20TransferStub.resolves(BigNumber.from(FAKE_GAS_COST));
-
-        const request: TokenGasEstimationParams = {
-          relayRequest: {
-            ...FAKE_RELAY_REQUEST,
-            relayData: {
-              ...FAKE_RELAY_REQUEST.relayData,
-              gasPrice: FAKE_GAS_COST,
-            }
-          },
-          internalEstimationCorrection: INTERNAL_CORRECTION,
-          estimatedGasCorrectionFactor: CORRECTION_FACTOR,
-        };
-
-        const estimation = await relayClient.estimateTokenTransferGas(request);
-
-        expect(estimation.toString()).to.equal(EXPECTED_ESTIMATION.toString());
-      });
-
-      it('should use by-default values when not sent as parameters', async function () {
-        //Just to be sure that the gas cost is lower than the estimate correction
-        const FAKE_GAS_COST = INTERNAL_TRANSACTION_ESTIMATED_CORRECTION - 1;
-        const EXPECTED_ESTIMATION = FAKE_GAS_COST;
-
-        ierc20TransferStub.resolves(BigNumber.from(FAKE_GAS_COST));
-
-        const request: TokenGasEstimationParams = {
-          relayRequest: {
-            ...FAKE_RELAY_REQUEST,
-            relayData: {
-              ...FAKE_RELAY_REQUEST.relayData,
-              gasPrice: FAKE_GAS_COST,
-            }
-          }
-        };
-
-        const estimation = await relayClient.estimateTokenTransferGas(request);
-
-        expect(estimation.toString()).to.equal(EXPECTED_ESTIMATION.toString());
-      });
-    });
-
-    describe('getSmartWalletAddress', function(){
-
-      let owner: string;
-      let index: number;
-
-      beforeEach(function () {
-        owner = createRandomAddress();
-        index = createRandomBigNumber(1000).toNumber();
-        const fakeConfig = {
-          config: {},
-          bytecodeHash: ''
-        } as discoveryUtils.Setup
-        sandbox.stub(discoveryUtils, 'setupDiscovery').returns(Promise.resolve(fakeConfig));
-      });
-
-      afterEach(function () {
-        sandbox.restore();
-      });
-
-      it('should return SW address', async function(){
-        const expectedAddress = createRandomAddress();
-        sandbox.stub(discoveryUtils, 'getSWAddress').returns(expectedAddress);
-        const address = await relayClient.getSmartWalletAddress(owner, index);
-
-        expect(address).to.be.equals(expectedAddress);
-      })
-
-    });
-
+  
     describe('relayTransaction', function () {
 
       const relayInfo: RelayInfo = {
@@ -1239,12 +870,12 @@ describe('RelayClient', function () {
       let attemptRelayTransactionStub: SinonStub;
       
       beforeEach(function () {
-        attemptRelayTransactionStub = sandbox.stub().returns(Promise.resolve(transaction));
+        attemptRelayTransactionStub = sandbox.stub().resolves(transaction);
         relayClient._getEnvelopingRequestDetails = sandbox.stub().returns(FAKE_RELAY_REQUEST);
         relayClient._prepareHttpRequest = sandbox.stub();
-        relayClient._verifyEnvelopingRequest = sandbox.stub().returns(Promise.resolve(true));
+        relayClient._verifyEnvelopingRequest = sandbox.stub().resolves(true);
         relayClient._attemptRelayTransaction = attemptRelayTransactionStub;
-        selectNextRelayStub = sandbox.stub(relayUtils, 'selectNextRelay').returns(Promise.resolve(relayInfo));
+        selectNextRelayStub = sandbox.stub(relayUtils, 'selectNextRelay').resolves(relayInfo);
       });
 
       afterEach(function () {
@@ -1252,16 +883,16 @@ describe('RelayClient', function () {
       })
 
       it('should relay the transaction in the first attempt', async function () {
-        const expectedTransaction = await relayClient.relayTransaction(envelopingRequest, FAKE_REQUEST_CONFIG);
+        const expectedTransaction = await relayClient.relayTransaction(envelopingRequest);
 
         expect(selectNextRelayStub).to.be.calledOnce;
         expect(expectedTransaction).to.be.equals(transaction);
       });
 
       it('should relay the transaction after the first attempt', async function () {
-        attemptRelayTransactionStub.onFirstCall().returns(Promise.resolve(undefined));
-        attemptRelayTransactionStub.onSecondCall().returns(Promise.resolve(transaction));
-        const expectedTransaction = await relayClient.relayTransaction(envelopingRequest, FAKE_REQUEST_CONFIG);
+        attemptRelayTransactionStub.onFirstCall().resolves(undefined);
+        attemptRelayTransactionStub.onSecondCall().resolves(transaction);
+        const expectedTransaction = await relayClient.relayTransaction(envelopingRequest);
 
         expect(selectNextRelayStub).to.be.calledTwice;
         expect(expectedTransaction).to.be.equals(transaction);
@@ -1270,7 +901,7 @@ describe('RelayClient', function () {
       
       it('should fail to relay transaction if details are missing', async function () {
         relayClient._getEnvelopingRequestDetails = sandbox.stub().throws(new Error('missing details'));
-        const expectedTransaction = relayClient.relayTransaction(envelopingRequest, FAKE_REQUEST_CONFIG);
+        const expectedTransaction = relayClient.relayTransaction(envelopingRequest);
 
         await expect(expectedTransaction).to.be.rejectedWith('missing details')
       });
@@ -1278,24 +909,24 @@ describe('RelayClient', function () {
       it('should fail to relay transaction if http request cannot be prepared', async function () {
         const error = new Error('FeesReceiver has to be a valid non-zero address');
         relayClient._prepareHttpRequest = sandbox.stub().throws(error);
-        const expectedTransaction = relayClient.relayTransaction(envelopingRequest, FAKE_REQUEST_CONFIG);
+        const expectedTransaction = relayClient.relayTransaction(envelopingRequest);
 
         await expect(expectedTransaction).to.be.rejectedWith(error.message)
       });
       
 
       it('should fail to relay transaction if there is no available relay server', async function () {
-        selectNextRelayStub.returns(Promise.resolve(undefined));
-        const expectedTransaction = relayClient.relayTransaction(envelopingRequest, FAKE_REQUEST_CONFIG);
+        selectNextRelayStub.resolves(undefined);
+        const expectedTransaction = relayClient.relayTransaction(envelopingRequest);
 
         await expect(expectedTransaction).to.be.rejectedWith(NOT_RELAYED_TRANSACTION)
       });
 
       it('should fail to relay transaction if cannot be verified locally on all servers', async function () {
-        relayClient._verifyEnvelopingRequest = sandbox.stub().returns(Promise.resolve(false));
-        selectNextRelayStub.onFirstCall().returns(Promise.resolve(relayInfo));
-        selectNextRelayStub.onSecondCall().returns(Promise.resolve(undefined));
-        const expectedTransaction = relayClient.relayTransaction(envelopingRequest, FAKE_REQUEST_CONFIG);
+        relayClient._verifyEnvelopingRequest = sandbox.stub().resolves(false);
+        selectNextRelayStub.onFirstCall().resolves(relayInfo);
+        selectNextRelayStub.onSecondCall().resolves(undefined);
+        const expectedTransaction = relayClient.relayTransaction(envelopingRequest);
 
         await expect(expectedTransaction).to.be.rejectedWith(NOT_RELAYED_TRANSACTION)
       });
@@ -1321,7 +952,7 @@ describe('RelayClient', function () {
         relayClient._httpClient = httpClient;
         parseTransactionStub = sandbox.stub(etherUtils, 'parse').returns(transaction);
         validateResponseStub = sandbox.stub(relayUtils, 'validateRelayResponse').returns(undefined);
-        relayClient._broadcastTx = sandbox.stub().returns(Promise.resolve(undefined));
+        relayClient._broadcastTx = sandbox.stub().resolves(undefined);
       });
 
       afterEach(function () {
@@ -1334,7 +965,7 @@ describe('RelayClient', function () {
         expect(expectedTransaction).to.be.equals(transaction);
       });
 
-      it('shoul throw if server cannot relay transaction', async function () {
+      it('should throw if server cannot relay transaction', async function () {
         httpClient.relayTransaction.throws('Got invalid response from relay: signedTx field missing.');
         const attempRelay = await relayClient._attemptRelayTransaction(relayInfo, FAKE_RELAY_TRANSACTION_REQUEST);
 
@@ -1376,7 +1007,7 @@ describe('RelayClient', function () {
 
       beforeEach(function () {
         sendTransactionStub = sandbox.stub();
-        relayClient._provider.sendTransaction = sendTransactionStub;
+        providerStub.sendTransaction = sendTransactionStub;
       });
 
       afterEach(function () {
@@ -1384,8 +1015,8 @@ describe('RelayClient', function () {
       })
 
       it('should send transaction if cannot find it in pool or receipt', async function () {
-        relayClient._provider.getTransaction = sandbox.stub().returns(undefined);
-        relayClient._provider.getTransactionReceipt = sandbox.stub().returns(undefined);
+        providerStub.getTransaction = sandbox.stub().returns(undefined);
+        providerStub.getTransactionReceipt = sandbox.stub().returns(undefined);
         await relayClient._broadcastTx(fakeSignedTx);
 
         expect(sendTransactionStub.called).to.be.true;
@@ -1393,8 +1024,8 @@ describe('RelayClient', function () {
 
       it('should not send transaction if the receipt is in the transaction pool', async function () {
         const fakeResponse = {} as TransactionResponse;
-        relayClient._provider.getTransaction = sandbox.stub().returns(Promise.resolve(fakeResponse));
-        relayClient._provider.getTransactionReceipt = sandbox.stub().returns(undefined);
+        providerStub.getTransaction = sandbox.stub().returns(Promise.resolve(fakeResponse));
+        providerStub.getTransactionReceipt = sandbox.stub().returns(undefined);
         await relayClient._broadcastTx(fakeSignedTx);
 
         expect(sendTransactionStub.called).to.be.false;
@@ -1402,8 +1033,8 @@ describe('RelayClient', function () {
 
       it('should not send transaction if receipt is found', async function () {
         const fakeReceipt = {} as TransactionReceipt;
-        relayClient._provider.getTransactionReceipt = sandbox.stub().returns(Promise.resolve(fakeReceipt));
-        relayClient._provider.getTransaction = sandbox.stub().returns(undefined);
+        providerStub.getTransactionReceipt = sandbox.stub().returns(Promise.resolve(fakeReceipt));
+        providerStub.getTransaction = sandbox.stub().returns(undefined);
         await relayClient._broadcastTx(fakeSignedTx);
 
         expect(sendTransactionStub.called).to.be.false;
@@ -1477,14 +1108,14 @@ describe('RelayClient', function () {
       });
 
       it('should succeed if worker has enough balance', async function () {
-        relayClient._provider.getBalance = sandbox.stub().returns(BigNumber.from(1_500_000));
+        providerStub.getBalance = sandbox.stub().returns(BigNumber.from(1_500_000));
         const verifyWorkerBalance = relayClient._verifyWorkerBalance(relayWorkerAddress, fakeMaxPossibleGas, FAKE_GAS_PRICE);
 
         await expect(verifyWorkerBalance).to.be.fulfilled;
       });
 
       it('should throw if worker has not enough balance', async function () {
-        relayClient._provider.getBalance = sandbox.stub().returns(BigNumber.from(500_000));
+        providerStub.getBalance = sandbox.stub().returns(BigNumber.from(500_000));
         const verifyWorkerBalance = relayClient._verifyWorkerBalance(relayWorkerAddress, fakeMaxPossibleGas, FAKE_GAS_PRICE);
 
         await expect(verifyWorkerBalance).to.be.rejectedWith('Worker does not have enough balance to pay');
