+++ test/gasEstimator/gasEstimator.test.ts
@@ -21,11 +21,11 @@ import {
   standardMaxPossibleGasEstimation,
 } from '../../src/gasEstimator/utils';
 import { ESTIMATED_GAS_CORRECTION_FACTOR } from '../../src/utils';
-import RelayClient from '../../src/RelayClient';
 
+import { createRandomAddress } from '../utils';
 import * as gasEstimatorUtils from '../../src/gasEstimator/utils';
 import * as clientConfiguration from '../../src/common/clientConfigurator';
-import { createRandomAddress } from '../utils';
+import * as relayUtils from '../../src/utils';
 
 use(sinonChai);
 use(chaiAsPromised);
@@ -54,10 +54,10 @@ describe('GasEstimator', function () {
       fakeFitRelayEstimation = randomBigNumber(10000);
       relayWorker = Wallet.createRandom().address;
       sandbox
-        .stub(RelayClient.prototype, 'estimateTokenTransferGas')
+        .stub(relayUtils, 'estimateTokenTransferGas')
         .returns(Promise.resolve(fakeTokenGas));
       sandbox
-        .stub(RelayClient.prototype, 'getSmartWalletAddress')
+        .stub(relayUtils, 'getSmartWalletAddress')
         .returns(Promise.resolve(createRandomAddress()));
     });
 
@@ -229,7 +229,7 @@ describe('GasEstimator', function () {
       const internalGas = randomBigNumber(10000);
 
       sandbox
-        .stub(RelayClient.prototype, 'estimateInternalCallGas')
+        .stub(relayUtils, 'estimateInternalCallGas')
         .returns(Promise.resolve(internalGas));
 
       const estimation = await linearFitMaxPossibleGasEstimation(
